Arrange, rename and clean up files
