--- legacy/app/activity/activity-lazy.js
@@ -1,3 +1,0 @@
-angular.module('ushahidi.routes', [])
-
-.config(require('./activity-routes.js'));
